+++ testing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # testing
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.7.1
+  - @bs-plugins/jira@1.0.9
+  - @bs-plugins/template@0.0.9
+
 ## 0.0.7
 
 ### Patch Changes